requirements.txt

numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.66
